Fix compiler warnings

+++ src/apps/S3DAnalyzer/utilities/depthintensitygradient.cpp
@@ -4,7 +4,7 @@ DepthIntensityGradient::DepthIntensityGradient(const QPointF& topLeft, const QPo
     : QLinearGradient(topLeft, bottomRight),
       colors{red, orange, green, green, blue, purple},
       steps{0.0f, 0.1f, 0.2f, 0.8f, 0.9f, 1.0f} {
-  for (int i = 0; i < colors.size(); ++i) {
+  for (size_t i = 0; i < colors.size(); ++i) {
     setColorAt(steps[i], colors[i]);
   }
 }
